Move the help command it it's own class. Also fix the help command to work with intrinsic commands.
