+++ test/Microsoft.DotNet.Tools.Tests.Utilities/Commands/HelpCommand.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.DotNet.Cli.Utils;
+
+namespace Microsoft.DotNet.Tools.Test.Utilities
+{
+    public sealed class HelpCommand : TestCommand
+    {
+        public HelpCommand()
+            : base("dotnet")
+        {
+        }
+
+        public override CommandResult Execute(string args = "")
+        {
+            args = $"help {args}";
+            return base.Execute(args);
+        }
+    }
+}
